pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Product deduplication challenge"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphframes>=0.6",
    "numpy>=2.3.3",
    "pandas==2.3.2",
    "psycopg2==2.9.10",
    "pyspark==3.4.1",
    "python-dotenv>=1.1.1",
    "setuptools==80.9.0",
    "sqlalchemy==2.0.43",
]
